pyproject.toml
[tool.poetry]
name = "ai_writer"
version = "0.1.0"
description = ""
authors = ["Aleksandr Kariakin <aleksandr.kariakin@toptal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.0"
langgraph = "^0.2.27"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.0"
duckduckgo-search = "^6.2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
